Locked bodies and drawables
Defined a locked body with a drawable super class. I'm not keen on this
arrangement but it will do until I get a render class or similar built.

Added world to screen. Makes sense that 0,0 world space should match the
center of my view when we are building a solar system like scene